# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Документация к сервису",
        "title": "API документация",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Isfandiyor",
            "email": "isfand.zabirov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "paths": {
        "/api/v1/notifications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Getting all notification list",
                "operationId": "get-notifications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notifications.Notification"
                            }
                        }
                    },
                    "400": {
                        "description": "reason",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Creating notification",
                "operationId": "create-notification",
                "parameters": [
                    {
                        "description": "The type of operation. Can be Purchase, VerifyCard or SendOtp",
                        "name": "orderType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "The session id of the operation",
                        "name": "sessionId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "The card pan",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "The date of the operation",
                        "name": "eventDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "The source of operation",
                        "name": "websiteUrl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Event data",
                        "name": "model",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/notifications.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "reason",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "notifications.Event": {
            "type": "object",
            "properties": {
                "card": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "orderType": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "websiteUrl": {
                    "type": "string"
                }
            }
        },
        "notifications.Notification": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sent": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}